Revert "refactor: split into multiple modules (#196)"
This reverts commit 737d5b1181264a51c56f5e0776154bca14d945ea.
