+++ .github/workflows/ci.yml
@@ -10,8 +10,8 @@ concurrency:
   group: ${{ github.ref }}-${{ github.workflow }}
   cancel-in-progress: true
 jobs:
-  commit:
-    name: Commits
+  conventional_commit:
+    name: Conventional Commits
     if: ${{ github.ref != 'refs/heads/main' }}
     runs-on: ubuntu-latest
     steps:
